Ensure apache default site has section for port 443
The apache default site was meant to always return 404. However, when
the wildcard sites functionality was added, the default site
configuration file was changed and it did not have a virtual host for
port 443 unless wildcard sites were configured. Thus, for https, another
configuration file (instead of 000-default.conf) would act as the
default. This was fixed.